+++ src/ffi.rs
@@ -3,8 +3,8 @@
 #![allow(non_snake_case)]
 #![allow(dead_code)]
 
-#[cfg(target_os = "macos")]
+#[cfg(enable_avf)]
 include!(concat!(env!("OUT_DIR"), "/bindings_macos.rs"));
 
-#[cfg(target_os = "linux")]
+#[cfg(any(enable_v4l2, enable_pipewire))]
 include!(concat!(env!("OUT_DIR"), "/bindings_linux.rs"));